Changed many `int`s to `size_t`s

+++ Problem_5/problem_5.c
@@ -9,31 +9,32 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-void Eratosthenes_Sieve( int, int[] );
-void count_multiplicity( int, int[], int[] );
+void Eratosthenes_Sieve( size_t, size_t[] );
+void count_multiplicity( size_t, size_t[], size_t[] );
 
 
 int main( int argc, char **argv ) {
 	char *ptr;  // Pointer for use in strtol
 	// Get the limit from the first command line argument
-	int limit = strtol( argv[1], &ptr, 10 );
+	size_t limit = strtol( argv[1], &ptr, 10 );
 	// Initialize list to hold primes
-	int prime_list[limit];
-	int multiplicity[limit];
+	size_t prime_list[limit];
+	size_t multiplicity[limit];
 
 	Eratosthenes_Sieve( limit, prime_list );
 	count_multiplicity( limit, prime_list, multiplicity );
 
-	for( int i = 0; i <= limit; i++ ) {
+	// Print out the prime number factors and their multiplicities
+	for( size_t i = 0; i <= limit; i++ ) {
 		if( prime_list[i] == 1 && multiplicity[i] != 0 )
-			printf("Number: %d\tMultiplicity: %d\n", i, multiplicity[i]);
+			printf("Number: %ld\tMultiplicity: %ld\n", i, multiplicity[i]);
 	}
 
 	return 0;
 }
 
 
-void Eratosthenes_Sieve( int limit, int prime_list[] ) {
+void Eratosthenes_Sieve( size_t limit, size_t prime_list[] ) {
 	// This function sets each element in `prime_list` to be either 0 or 1,
 	// composite or prime, which the index is the number is question.
 
@@ -41,15 +42,15 @@ void Eratosthenes_Sieve( int limit, int prime_list[] ) {
 	prime_list[0] = 0;
 	prime_list[1] = 0;
 	// Initialize the rest of the list to be prime
-	for( int i = 2; i <= limit; i++ ) {
+	for( size_t i = 2; i <= limit; i++ ) {
 		prime_list[i] = 1;
 	}
 
 	// Loop to mark off the values that are not prime, according to the method
-	int p = 2;
+	size_t p = 2;
 	while( (p*p) <= limit ) {
 		if( prime_list[p] == 1 ) {
-			for( int i = (p*p); i <= limit; i += p ) {
+			for( size_t i = (p*p); i <= limit; i += p ) {
 				prime_list[i] = 0;
 			}
 		}
@@ -58,9 +59,12 @@ void Eratosthenes_Sieve( int limit, int prime_list[] ) {
 }
 
 
-void count_multiplicity( int target, int prime_list[], int multiplicity[] ) {
+void count_multiplicity( size_t target, size_t prime_list[], size_t multiplicity[] ) {
+	// Modifies the `multiplicity` array to house the multiplicity of a number
+	// number's prime factorization.
+
 	// Initialize multiplicity array to zeros
-	for( int i = 0; i <= target; i++ )
+	for( size_t i = 0; i <= target; i++ )
 		multiplicity[i] = 0;
 
 	// Boolean to be used when the target itself is prime
@@ -77,7 +81,7 @@ void count_multiplicity( int target, int prime_list[], int multiplicity[] ) {
 		last_prime_bool = 1;
 	}
 
-	for( int j = 3; j <= target; j += 2 ) {
+	for( size_t j = 3; j <= target; j += 2 ) {
 		while( target % j == 0 ) {
 			multiplicity[j] += 1;
 			target /= j;
